requirements.txt
accelerate==0.34.2
av==11.0.0
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
decord==0.6.0
diffusers==0.32.2
distributed @ file:///rapids/distributed-2024.1.1-py3-none-any.whl#sha256=cf05d3b38e1700339b3e36395729ab62110e723efefaecc21a8260fdc7555cf9
einops==0.4.1
ffmpy==0.4.0
gradio==3.41.2
gradio_client==0.5.0
imageio==2.33.0
imageio-ffmpeg==0.4.9
matplotlib==3.8.4
matplotlib-inline==0.1.7
mediapipe==0.10.8
numpy==1.23.5
omegaconf==2.2.3
onnx @ file:///opt/pytorch/pytorch/third_party/onnx
onnxruntime==1.19.2
onnxruntime-gpu==1.16.3
open-clip-torch==2.20.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
pandas @ file:///rapids/pandas-2.2.1-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=c38ce92cb22a4bea4e3929429aa1067a454dcc9c335799af93ba9be21b6beb51
Pillow @ file:///tmp/pillow-simd
PyYAML==6.0.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.4
tensorboard==2.9.0
torch==2.6.0
torch-fidelity==0.3.0
torchmetrics==1.2.1
tqdm==4.66.1
transformers==4.43.2
xformers==0.0.29.post3
yacs==0.1.8